requirements.txt
Flask==3.0.0
pytest==8.2.2
google-generativeai==0.5.4
python-dotenv==1.0.1
gunicorn==22.0.0
pytest-cov==5.0.0
google-cloud-logging==3.10.0
google-cloud-storage==2.16.0
google-cloud-error-reporting==1.11.0
google-cloud-pubsub==2.21.0
firebase-admin==6.4.0
black==24.4.2
python-json-logger==2.0.7
